Moved SmartDerivatives to an independent module
