+++ mlcolvar/core/loss/__init__.py
@@ -24,5 +24,6 @@ from .eigvals import ReduceEigenvaluesLoss, reduce_eigenvalues_loss
 from .elbo import ELBOGaussiansLoss, elbo_gaussians_loss
 from .autocorrelation import AutocorrelationLoss, autocorrelation_loss
 from .fisher import FisherDiscriminantLoss, fisher_discriminant_loss
-from .committor_loss import CommittorLoss, committor_loss, SmartDerivatives, compute_descriptors_derivatives
+from .committor_loss import CommittorLoss, committor_loss
 from .generator_loss import GeneratorLoss
+from .utils.smart_derivatives import SmartDerivatives, compute_descriptors_derivatives
